requirements.txt
tensorflow
scikit-learn
pandas
numpy
matplotlib
streamlit
Pillow

